+++ R/cancensus_cache.R
@@ -1,4 +1,11 @@
-touch_metadata <- function(meta_file){
+metadata_columns <- c("path" , "dataset", "regions", "level", "vectors", "created_at",
+                      "version", "size" ,"last_accessed", "access_count",  "resolution")
+metadata_columns_c <- c("path" , "dataset", "regions", "level", "vectors", "version", "resolution")
+metadata_columns_d <- c("created_at","last_accessed")
+metadata_columns_n <- c("size" , "access_count")
+
+
+generate_metadata <- function(meta_file,params){
   if (file.exists(meta_file)) {
     metadata <- readRDS(meta_file) %>%
       dplyr::mutate(last_accessed=Sys.time())
@@ -10,8 +17,44 @@ touch_metadata <- function(meta_file){
   } else {
     pattern <- basename(meta_file) %>% gsub("\\.meta$","",.)
     path <- dir(cache_path(),pattern = pattern)
-    metadata <- dplyr::tibble(path=path, last_accessed=Sys.time(),access_count=1)
+    metadata <- dplyr::tibble(last_accessed=Sys.time(),access_count=1)
   }
+  if (!is.null(params) & !("dataset" %in% names(metadata)) & "dataset" %in% names(params)) {
+    metadata$dataset=params$dataset
+  }
+  if (!is.null(params) & !("level" %in% names(metadata)) & "level" %in% names(params)) {
+    metadata$level=params$level
+  }
+  if (!is.null(params) & !("regions" %in% names(metadata)) & "regions" %in% names(params)) {
+    metadata$regions=as.character(params$regions)
+  }
+  if (!is.null(params) & !("resolution" %in% names(metadata)) & "resolution" %in% names(params)) {
+    metadata$resolution=as.character(params$resolution)
+  }
+  if (!is.null(params) & !("vectors" %in% names(metadata)) & "vectors" %in% names(params)) {
+    metadata$vectors=jsonlite::toJSON(as.character(params$vectors))  %>% as.character()
+  }
+  if (!("version" %in% names(metadata))) {
+    path <- basename(meta_file)
+    if (grepl("^CM_geo_",path)) metadata$version="g.1"
+    if (grepl("^CM_data_",path)) metadata$version="d.1"
+  }
+
+  for (n in setdiff(metadata_columns_c,names(metadata))) {
+    metadata[,n]=NA_character_
+  }
+  # for (n in setdiff(metadata_columns_d,names(metadata))) {
+  #   metadata[,n]=NA
+  # }
+  for (n in setdiff(metadata_columns_n,names(metadata))) {
+    metadata[,n]=NA_real_
+  }
+
+  metadata
+}
+
+touch_metadata <- function(meta_file,params=NULL){
+  metadata <- generate_metadata(meta_file,params)
   saveRDS(metadata,file = meta_file)
 }
 
@@ -58,6 +101,18 @@ list_cancensus_cache <- function(){
       dplyr::mutate(size=dplyr::coalesce(.data$size,.data$manual_size)) %>%
       dplyr::select(-.data$manual_size)
   }
+
+  for (n in setdiff(metadata_columns_c,names(all))) {
+    all[,n]=NA_character_
+  }
+  # for (n in setdiff(metadata_columns_d,names(metadata))) {
+  #   metadata[,n]=NA
+  # }
+  for (n in setdiff(metadata_columns_n,names(all))) {
+    all[,n]=NA_real_
+  }
+
+
   all
 }