+++ vignettes/Making_maps_with_cancensus.R
@@ -1,9 +1,0 @@
-## ----setup, include = FALSE---------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  eval = nzchar(Sys.getenv("COMPILE_VIG"))
-)
-
-mb_token <- Sys.getenv("MB_TOKEN")
-